@@ -5,67 +5,60 @@
   "main": "bin/run",
   "dependencies": {
     "@types/bcrypt": "^1.0.0",
-    "@types/chalk": "^0.4.31",
-    "@types/handlebars": "^4.0.33",
-    "@types/koa-basic-auth": "^2.0.1",
-    "@types/koa-bodyparser": "^3.0.23",
-    "@types/koa-json": "^2.0.16",
-    "@types/koa-logger": "^2.0.2",
-    "@types/koa-mount": "^2.0.18",
-    "@types/koa-router": "^7.0.22",
+    "@types/dotenv": "^4.0.2",
+    "@types/koa-bodyparser": "^3.0.26",
+    "@types/koa-json": "^2.0.18",
+    "@types/koa-router": "^7.0.25",
     "@types/ms": "^0.7.29",
-    "@types/node": "^8.0.4",
+    "@types/node": "^8.0.53",
     "@types/pino": "^3.0.1",
     "@types/reflect-metadata": "^0.0.5",
-    "@types/request-promise-native": "^1.0.5",
+    "@types/request-promise-native": "^1.0.10",
     "@types/uuid": "^3.0.0",
     "bcrypt": "^1.0.2",
-    "config-file-bi": "^0.2.0",
-    "flexblocks": "^1.0.1",
-    "fs-extra": "^3.0.1",
-    "handlebars": "^4.0.10",
-    "koa": "^2.3.0",
-    "koa-basic-auth": "^2.0.0",
+    "cerialize": "^0.1.18",
+    "dotenv": "^4.0.0",
+    "koa": "^2.4.1",
     "koa-bodyparser": "^4.2.0",
     "koa-json": "^2.0.2",
-    "koa-logger": "^3.0.0",
-    "koa-mount": "^3.0.0",
-    "koa-router": "^7.2.1",
-    "material-design-lite": "^1.3.0",
+    "koa-router": "^7.3.0",
     "ms": "^2.0.0",
-    "mysql": "^2.13.0",
-    "pg": "^6.4.0",
-    "pino": "^4.6.0",
+    "pg": "^7.4.0",
+    "pino": "^4.10.2",
     "reflect-metadata": "^0.1.10",
     "request": "^2.81.0",
     "request-promise-native": "^1.0.4",
     "sendgrid": "^5.1.1",
-    "typeorm": "^0.1.0-alpha.22",
-    "typescript": "^2.4.1",
+    "tslint-language-service": "^0.9.6",
+    "typeorm": "^0.1.3",
+    "typescript": "^2.6.1",
     "uuid": "^3.1.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.0.1",
-    "@types/mocha": "^2.2.41",
+    "@types/chai": "^4.0.5",
+    "@types/mocha": "^2.2.44",
     "chai": "^4.0.2",
-    "coveralls": "^2.13.1",
-    "mocha": "^3.4.2",
+    "coveralls": "^3.0.0",
+    "mocha": "^4.0.1",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^11.0.3",
+    "nyc": "^11.3.0",
     "supervisor": "^0.12.0",
-    "tslint": "^5.4.3",
-    "tslint-consistent-codestyle": "^1.4.0",
+    "tslint": "^5.8.0",
+    "tslint-consistent-codestyle": "^1.10.0",
     "tslint-eslint-rules": "^4.1.1"
   },
   "scripts": {
-    "test": "nyc -r html mocha",
-    "test-cov": "nyc -r lcovonly -r html mocha",
+    "compile": "tsc && tslint -p . -e node_modules/**.* && node bin/buildOrmConfig",
+    "test": "nyc -r html mocha test/index",
+    "test-cov": "nyc -r lcovonly -r html mocha test/index",
     "postinstall": "tsc",
     "pretest": "tsc&& tslint -p . -e node_modules/**.*",
     "dev": "supervisor -i coverage bin/run | pino",
     "start": "node ./bin/run",
     "lint": "tslint -p . -e node_modules/**.*",
-    "test-nocompile": "nyc -r html mocha"
+    "test-nocompile": "nyc -r html mocha test/index",
+    "db:migrate": "node bin/buildOrmConfig && typeorm migrations:run",
+    "db:revert": "node bin/buildOrmConfig && typeorm migrations:revert"
   },
   "greenkeeper": {
     "ignore": []
