@@ -2,48 +2,35 @@
 
 import Router = require("koa-router");
 const router = new Router();
-import config from "../lib/config";
-import { User, Session } from "../models";
-import { connection as db } from "../lib/db";
-import log from "../lib/log";
+import { Session } from "../models";
 import * as Errors from "../lib/errors";
-import { authUser } from "../lib/auth";
-
-interface ICtxState {
-  authType: "Basic" | "Bearer";
-  user: User;
-  session?: Session;
-  // TODO: support confirmation& service
-}
+import { authBasic, authBearer } from "../lib/auth";
 
 router.put("/session", async (ctx) => {
-  await authUser(ctx, "Basic");
-  const state = ctx.state as ICtxState;
-  const session = new Session(state.user);
+  const user = await authBasic(ctx);
+  const session = new Session(user);
   if (ctx.request.body.permissions && ctx.request.body.permissions.admin) {
-    if (state.user.permissions.admin) {
+    if (user.permissions.admin) {
       session.permissions.admin = true;
     } else {
       throw new Errors.InsufficientPermissionError(session, "admin");
     }
   }
-  await db.getRepository(Session).persist(session);
+  await session.save();
   ctx.body = await session.toView();
 });
 router.get("/session", async (ctx) => {
-  await authUser(ctx, "Bearer");
-  const state = ctx.state as ICtxState;
-  if (state.session) {
-    ctx.body = await state.session.toView();
+  const session = await authBearer(ctx);
+  if (session) {
+    ctx.body = await session.toView();
   }
 });
 router.delete("/session", async (ctx) => {
-  await authUser(ctx, "Bearer");
-  const state = ctx.state as ICtxState;
-  if (state.session) {
-    state.session.expiresAt = new Date(Date.now());
-    await db.getRepository(Session).persist(state.session);
-    ctx.body = await state.session.toView(true);
+  const session = await authBearer(ctx);
+  if (session) {
+    session.expiresAt = new Date(Date.now());
+    await session.save();
+    ctx.body = await session.toView(true);
   }
 });
 
